refactor(cli): decompose SetupCommand and PiLoginLaunch flow modules
Delete the /* eslint-disable max-lines */ and complexity escape hatches from
SetupCommand.ts and PiLoginLaunch.ts and make both conform to the repo rules
(complexity 10, max-lines 500) by extracting cohesive flow modules:

- setup/SetupTypes.ts: shared setup types and prompt-choice constants
  (SetupCommand re-exports them, so RunCommand and existing imports are
  untouched).
- setup/SetupStarterLayout.ts: setup-source layout resolution, git sync,
  starter settings reading/validation.
- setup/SetupSourceImport.ts: copy/symlink import application, profile file
  copying, and default profile creation.
- setup/SetupPrompts.ts: interactive prompt flows (readline import
  target/mode/profile prompts, post-import launch question).
- setup/SetupProfileDiscovery.ts: profile choice discovery, source
  materialization, and launch-resolution guards.
- setup/SetupMessages.ts: user-facing setup summary and exit messages.
- PiLoginLaunch extracts the extension runtime-config builder so the launch
  planner stays under the complexity limit.

No behavior change: the existing suite passes unchanged and the packaged
e2e smoke stays green. v8 ignore pragmas in the setup code drop from 25 to
17 because the extracted readline prompts, symlink import, starter layout,
and exit-message paths now have direct unit tests (setup-modules.test.ts).

Guard the previously unguarded symlinkSync in symlinkLocalOutfitterSource:
on win32, EPERM/EACCES now produces an actionable error suggesting Windows
Developer Mode or the copy snapshot import mode instead of crashing (it
never silently copies), unit-tested with an injected throwing symlink.

RunCommand.ts is intentionally untouched (another agent owns it); its
eslint-disable header remains for a follow-up.

Closes #13
